requirements.txt
streamlit==1.42.0
youtube-transcript-api==1.2.4
google-genai==1.3.0
python-dotenv==1.0.1
pydantic==2.10.6
